Refactor: move things around
